# Conflicts:
#	src/test/java/com/epam/aidial/deployment/manager/service/deployment/InferenceDeploymentManagerTest.java
#	src/test/java/com/epam/aidial/deployment/manager/service/manifest/InferenceManifestGeneratorTest.java

+++ src/main/java/com/epam/aidial/deployment/manager/model/ScalingStrategy.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 public class ScalingStrategy {
     @JsonProperty("$type")
     private ScalingStrategyType type;
-    private double threshold;
+    private int threshold;
 }
